renamed file cand created class
